Changed ulong to int, added qsort

+++ Source/GenerateNumbers.cpp
@@ -4,34 +4,34 @@
 #include <algorithm>
 #include "GenerateNumbers.hpp"
 
-ulong randint(int min, int max)
+int randint(int min, int max)
 {
-    return min + (rand() % static_cast<ulong>(max - min + 1));
+    return min + (rand() % static_cast<int>(max - min + 1));
 }
 
-void saveNumbers(std::string filename, ulong numbers[], ulong size)
+void saveNumbers(std::string filename, int numbers[], int size)
 {
     std::ofstream file;
     file.open(filename);
-    for (ulong i = 0; i < size; ++i)
+    for (int i = 0; i < size; ++i)
     {
         file << numbers[i] << std::endl;
     }
     file.close();
 }
 
-ulong* generateNumbers()
+int* generateNumbers()
 {
-    ulong n = 10000000;
-    ulong* arr = new ulong[n];
+    int n = 10000000;
+    int* arr = new int[n];
 
-    for (ulong i = 0; i < n; ++i)
+    for (int i = 0; i < n; ++i)
     {
         arr[i] = i;
     }
-    for (ulong i = 0; i < n; ++i)
+    for (int i = 0; i < n; ++i)
     {
-        std::swap<ulong>(arr[i], arr[randint(0, n)]);
+        std::swap<int>(arr[i], arr[randint(0, n)]);
     }
     return arr;
 }
